# core/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/service-accounts": {
            "post": {
                "description": "Create a new IAM service account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Service Account",
                "operationId": "create-service-account",
                "parameters": [
                    {
                        "description": "Service account details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.CreateServiceAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/iam.ServiceAccount"
                        }
                    }
                }
            }
        },
        "/service-accounts/{project_id}": {
            "get": {
                "description": "Get a list of IAM service accounts",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Service Accounts",
                "operationId": "list-service-accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.ServiceAccountDetails"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Get a list of IAM service accounts",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Service Account",
                "operationId": "list-service-accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.ServiceAccountDetails"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.CreateServiceAccountRequest": {
            "type": "object",
            "required": [
                "description",
                "display_name",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_responses_gcp.ServiceAccountDetails": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "displayName": {
                    "type": "string"
                },
                "etag": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "uniqueId": {
                    "type": "string"
                }
            }
        },
...
        "iam.ServiceAccount": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description: Optional. A user-specified, human-readable description\nof the service account. The maximum length is 256 UTF-8 bytes.",
                    "type": "string"
                },
                "disabled": {
                    "description": "Disabled: Output only. Whether the service account is disabled.",
                    "type": "boolean"
                },
                "displayName": {
                    "description": "DisplayName: Optional. A user-specified, human-readable name for the\nservice account. The maximum length is 100 UTF-8 bytes.",
                    "type": "string"
                },
                "email": {
                    "description": "Email: Output only. The email address of the service account.",
                    "type": "string"
                },
                "etag": {
                    "description": "Etag: Deprecated. Do not use.",
                    "type": "string"
                },
                "name": {
                    "description": "Name: The resource name of the service account. Use one of the\nfollowing formats: *\n`projects/{PROJECT_ID}/serviceAccounts/{EMAIL_ADDRESS}` *\n`projects/{PROJECT_ID}/serviceAccounts/{UNIQUE_ID}` As an\nalternative, you can use the `-` wildcard character instead of the\nproject ID: * `projects/-/serviceAccounts/{EMAIL_ADDRESS}` *\n`projects/-/serviceAccounts/{UNIQUE_ID}` When possible, avoid using\nthe `-` wildcard character, because it can cause response messages to\ncontain misleading error codes. For example, if you try to access the\nservice account `projects/-/serviceAccounts/fake@example.com`, which\ndoes not exist, the response contains an HTTP `403 Forbidden` error\ninstead of a `404 Not Found` error.",
                    "type": "string"
                },
                "oauth2ClientId": {
                    "description": "Oauth2ClientId: Output only. The OAuth 2.0 client ID for the service\naccount.",
                    "type": "string"
                },
                "projectId": {
                    "description": "ProjectId: Output only. The ID of the project that owns the service\naccount.",
                    "type": "string"
                },
                "uniqueId": {
                    "description": "UniqueId: Output only. The unique, stable numeric ID for the service\naccount. Each service account retains its unique ID even if you\ndelete the service account. For example, if you delete a service\naccount, then create a new service account with the same name, the\nnew service account has a different unique ID than the deleted\nservice account.",
                    "type": "string"
                }
            }
        },
...
    }
}